changes to utilize the added sdk functionality
